requirements.txt
flask
flask-cors
openpyxl
pillow
gunicorn
psycopg2-binary
